Move functions from helpers/index.js into individual files
